QueueClient is flattened into QSC, MC is renamed to QC, MidC is flattened into QC
