+++ sdk/storage/storage-queue/src/QueueServiceClient.ts
@@ -1,21 +1,26 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { TokenCredential, isTokenCredential, isNode } from "@azure/core-http";
+import { TokenCredential, isTokenCredential, isNode, HttpResponse } from "@azure/core-http";
 import { CanonicalCode } from "@azure/core-tracing";
 import * as Models from "./generated/src/models";
 import { AbortSignalLike } from "@azure/abort-controller";
 import { ListQueuesIncludeType } from "./generated/src/models";
-import { Service } from "./generated/src/operations";
+import { Service, Queue } from "./generated/src/operations";
 import { newPipeline, NewPipelineOptions, Pipeline } from "./Pipeline";
 import { StorageClient, CommonOptions } from "./StorageClient";
-import { QueueClient } from "./QueueClient";
 import "@azure/core-paging";
 import { PageSettings, PagedAsyncIterableIterator } from "@azure/core-paging";
-import { appendToURLPath, extractConnectionStringParts } from "./utils/utils.common";
+import {
+  appendToURLPath,
+  extractConnectionStringParts,
+  truncatedISO8061Date
+} from "./utils/utils.common";
 import { SharedKeyCredential } from "./credentials/SharedKeyCredential";
 import { AnonymousCredential } from "./credentials/AnonymousCredential";
 import { createSpan } from "./utils/tracing";
+import { Metadata } from "./models";
+import { QueueClient } from "./QueueClient";
 
 /**
  * Options to configure Queue Service - Get Properties operation
@@ -133,6 +138,169 @@ export interface ServiceListQueuesOptions extends CommonOptions {
 }
 
 /**
+ * Options to configure Queue - Create operation
+ *
+ * @export
+ * @interface QueueCreateOptions
+ */
+export interface QueueCreateOptions extends CommonOptions {
+  /**
+   * An implementation of the `AbortSignalLike` interface to signal the request to cancel the operation.
+   * For example, use the &commat;azure/abort-controller to create an `AbortSignal`.
+   *
+   * @type {AbortSignalLike}
+   * @memberof AppendBlobCreateOptions
+   */
+  abortSignal?: AbortSignalLike;
+  /**
+   * A collection of key-value string pair to associate with the queue object.
+   * The keys need to be lower-case.
+   *
+   * @type {Metadata}
+   * @memberof QueueCreateOptions
+   */
+  metadata?: Metadata;
+}
+
+/**
+ * Options to configure Queue - Get Properties operation
+ *
+ * @export
+ * @interface QueueGetPropertiesOptions
+ */
+export interface QueueGetPropertiesOptions extends CommonOptions {
+  /**
+   * An implementation of the `AbortSignalLike` interface to signal the request to cancel the operation.
+   * For example, use the &commat;azure/abort-controller to create an `AbortSignal`.
+   *
+   * @type {AbortSignalLike}
+   * @memberof AppendBlobCreateOptions
+   */
+  abortSignal?: AbortSignalLike;
+}
+
+/**
+ * Options to configure Queue - Delete operation
+ *
+ * @export
+ * @interface QueueDeleteOptions
+ */
+export interface QueueDeleteOptions extends CommonOptions {
+  /**
+   * An implementation of the `AbortSignalLike` interface to signal the request to cancel the operation.
+   * For example, use the &commat;azure/abort-controller to create an `AbortSignal`.
+   *
+   * @type {AbortSignalLike}
+   * @memberof AppendBlobCreateOptions
+   */
+  abortSignal?: AbortSignalLike;
+}
+
+/**
+ * Options to configure Queue - Get Access Policy operation
+ *
+ * @export
+ * @interface QueueGetAccessPolicyOptions
+ */
+export interface QueueGetAccessPolicyOptions extends CommonOptions {
+  /**
+   * An implementation of the `AbortSignalLike` interface to signal the request to cancel the operation.
+   * For example, use the &commat;azure/abort-controller to create an `AbortSignal`.
+   *
+   * @type {AbortSignalLike}
+   * @memberof AppendBlobCreateOptions
+   */
+  abortSignal?: AbortSignalLike;
+}
+
+/**
+ * Options to configure Queue - Set Access Policy operation
+ *
+ * @export
+ * @interface QueueSetAccessPolicyOptions
+ */
+export interface QueueSetAccessPolicyOptions extends CommonOptions {
+  /**
+   * An implementation of the `AbortSignalLike` interface to signal the request to cancel the operation.
+   * For example, use the &commat;azure/abort-controller to create an `AbortSignal`.
+   *
+   * @type {AbortSignalLike}
+   * @memberof AppendBlobCreateOptions
+   */
+  abortSignal?: AbortSignalLike;
+}
+
+/**
+ * Options to configure Queue - Set Metadata operation
+ *
+ * @export
+ * @interface QueueSetMetadataOptions
+ */
+export interface QueueSetMetadataOptions extends CommonOptions {
+  /**
+   * An implementation of the `AbortSignalLike` interface to signal the request to cancel the operation.
+   * For example, use the &commat;azure/abort-controller to create an `AbortSignal`.
+   *
+   * @type {AbortSignalLike}
+   * @memberof AppendBlobCreateOptions
+   */
+  abortSignal?: AbortSignalLike;
+}
+
+/**
+ * Signed identifier.
+ *
+ * @export
+ * @interface SignedIdentifier
+ */
+export interface SignedIdentifier {
+  /**
+   * @member {string} id a unique id
+   */
+  id: string;
+  /**
+   * @member {AccessPolicy} accessPolicy
+   */
+  accessPolicy: {
+    /**
+     * @member {Date} start the date-time the policy is active.
+     */
+    start: Date;
+    /**
+     * @member {string} expiry the date-time the policy expires.
+     */
+    expiry: Date;
+    /**
+     * @member {string} permission the permissions for the acl policy
+     * @see https://docs.microsoft.com/en-us/rest/api/storageservices/set-queue-acl
+     */
+    permission: string;
+  };
+}
+
+export declare type QueueGetAccessPolicyResponse = {
+  signedIdentifiers: SignedIdentifier[];
+} & Models.QueueGetAccessPolicyHeaders & {
+    /**
+     * The underlying HTTP response.
+     */
+    _response: HttpResponse & {
+      /**
+       * The parsed HTTP response headers.
+       */
+      parsedHeaders: Models.QueueGetAccessPolicyHeaders;
+      /**
+       * The response body as text (string format)
+       */
+      bodyAsText: string;
+      /**
+       * The response body as parsed JSON or XML
+       */
+      parsedBody: Models.SignedIdentifier[];
+    };
+  };
+
+/**
  * A QueueServiceClient represents a URL to the Azure Storage Queue service allowing you
  * to manipulate queues.
  *
@@ -140,6 +308,27 @@ export interface ServiceListQueuesOptions extends CommonOptions {
  * @class QueueServiceClient
  */
 export class QueueServiceClient extends StorageClient {
+  /**
+   * serviceContext provided by protocol layer.
+   *
+   * @private
+   * @type {Service}
+   * @memberof QueueServiceClient
+   */
+  private serviceContext: Service;
+  /**
+   * queueContext provided by protocol layer.
+   *
+   * @private
+   * @type {Queue}
+   * @memberof QueueServiceClient
+   */
+  private queueContext: Queue;
+  private _queueName: string;
+  public get queueName(): string {
+    return this._queueName;
+  }
+
   /**
    * Creates an instance of QueueServiceClient.
    *
@@ -149,53 +338,18 @@ export class QueueServiceClient extends StorageClient {
    *                                  `DefaultEndpointsProtocol=https;AccountName=myaccount;AccountKey=accountKey;EndpointSuffix=core.windows.net`
    *                                  SAS connection string example -
    *                                  `BlobEndpoint=https://myaccount.blob.core.windows.net/;QueueEndpoint=https://myaccount.queue.core.windows.net/;FileEndpoint=https://myaccount.file.core.windows.net/;TableEndpoint=https://myaccount.table.core.windows.net/;SharedAccessSignature=sasString`
+   * @param {string} queueName Queue name.
    * @param {NewPipelineOptions} [options] Options to configure the HTTP pipeline.
-   * @returns {QueueServiceClient} A new QueueServiceClient object from the given connection string.
    * @memberof QueueServiceClient
    */
-  public static fromConnectionString(
-    connectionString: string,
-    options?: NewPipelineOptions
-  ): QueueServiceClient {
-    options = options || {};
-    const extractedCreds = extractConnectionStringParts(connectionString);
-    if (extractedCreds.kind === "AccountConnString") {
-      if (isNode) {
-        const sharedKeyCredential = new SharedKeyCredential(
-          extractedCreds.accountName!,
-          extractedCreds.accountKey
-        );
-        options.proxy = extractedCreds.proxyUri;
-        const pipeline = newPipeline(sharedKeyCredential, options);
-        return new QueueServiceClient(extractedCreds.url, pipeline);
-      } else {
-        throw new Error("Account connection string is only supported in Node.js environment");
-      }
-    } else if (extractedCreds.kind === "SASConnString") {
-      const pipeline = newPipeline(new AnonymousCredential(), options);
-      return new QueueServiceClient(extractedCreds.url + "?" + extractedCreds.accountSas, pipeline);
-    } else {
-      throw new Error(
-        "Connection string must be either an Account connection string or a SAS connection string"
-      );
-    }
-  }
-
-  /**
-   * serviceContext provided by protocol layer.
-   *
-   * @private
-   * @type {Service}
-   * @memberof QueueServiceClient
-   */
-  private serviceContext: Service;
-
+  constructor(connectionString: string, queueName: string, options?: NewPipelineOptions);
   /**
    * Creates an instance of QueueServiceClient.
    *
-   * @param {string} url A URL string pointing to Azure Storage queue service, such as
-   *                     "https://myaccount.queue.core.windows.net". You can append a SAS
-   *                     if using AnonymousCredential, such as "https://myaccount.queue.core.windows.net?sasString".
+   * @param {string} url A URL string pointing to Azure Storage queue, such as
+   *                     "https://myaccount.queue.core.windows.net/myqueue". You can
+   *                     append a SAS if using AnonymousCredential, such as
+   *                     "https://myaccount.queue.core.windows.net/myqueue?sasString".
    * @param {SharedKeyCredential | AnonymousCredential | TokenCredential} credential Such as AnonymousCredential, SharedKeyCredential
    *                                                  or a TokenCredential from @azure/identity. If not specified,
    *                                                  AnonymousCredential is used.
@@ -210,42 +364,83 @@ export class QueueServiceClient extends StorageClient {
   /**
    * Creates an instance of QueueServiceClient.
    *
-   * @param {string} url A URL string pointing to Azure Storage queue service, such as
-   *                     "https://myaccount.queue.core.windows.net". You can append a SAS
-   *                     if using AnonymousCredential, such as "https://myaccount.queue.core.windows.net?sasString".
+   * @param {string} url A URL string pointing to Azure Storage queue, such as
+   *                     "https://myaccount.queue.core.windows.net/myqueue". You can
+   *                     append a SAS if using AnonymousCredential, such as
+   *                     "https://myaccount.queue.core.windows.net/myqueue?sasString".
    * @param {Pipeline} pipeline Call newPipeline() to create a default
    *                            pipeline, or provide a customized pipeline.
    * @memberof QueueServiceClient
    */
   constructor(url: string, pipeline: Pipeline);
   constructor(
-    url: string,
-    credentialOrPipeline?: SharedKeyCredential | AnonymousCredential | TokenCredential | Pipeline,
+    urlOrConnectionString: string,
+    credentialOrPipelineOrQueueName?:
+      | SharedKeyCredential
+      | AnonymousCredential
+      | TokenCredential
+      | Pipeline
+      | string,
     options?: NewPipelineOptions
   ) {
+    options = options || {};
     let pipeline: Pipeline;
-    if (credentialOrPipeline instanceof Pipeline) {
-      pipeline = credentialOrPipeline;
+    let url: string;
+    if (credentialOrPipelineOrQueueName instanceof Pipeline) {
+      // (url: string, pipeline: Pipeline)
+      url = urlOrConnectionString;
+      pipeline = credentialOrPipelineOrQueueName;
     } else if (
-      (isNode && credentialOrPipeline instanceof SharedKeyCredential) ||
-      credentialOrPipeline instanceof AnonymousCredential ||
-      isTokenCredential(credentialOrPipeline)
+      (isNode && credentialOrPipelineOrQueueName instanceof SharedKeyCredential) ||
+      credentialOrPipelineOrQueueName instanceof AnonymousCredential ||
+      isTokenCredential(credentialOrPipelineOrQueueName)
     ) {
-      pipeline = newPipeline(credentialOrPipeline, options);
-    } else {
+      // (url: string, credential?: SharedKeyCredential | AnonymousCredential | TokenCredential, options?: NewPipelineOptions)
+      url = urlOrConnectionString;
+      pipeline = newPipeline(credentialOrPipelineOrQueueName, options);
+    } else if (
+      !credentialOrPipelineOrQueueName &&
+      typeof credentialOrPipelineOrQueueName !== "string"
+    ) {
+      // (url: string, credential?: SharedKeyCredential | AnonymousCredential | TokenCredential, options?: NewPipelineOptions)
       // The second paramter is undefined. Use anonymous credential.
+      url = urlOrConnectionString;
       pipeline = newPipeline(new AnonymousCredential(), options);
+    } else if (
+      credentialOrPipelineOrQueueName &&
+      typeof credentialOrPipelineOrQueueName === "string"
+    ) {
+      // (connectionString: string, containerName: string, queueName: string, options?: NewPipelineOptions)
+      const extractedCreds = extractConnectionStringParts(urlOrConnectionString);
+      if (extractedCreds.kind === "AccountConnString") {
+        if (isNode) {
+          const queueName = credentialOrPipelineOrQueueName;
+          const sharedKeyCredential = new SharedKeyCredential(
+            extractedCreds.accountName,
+            extractedCreds.accountKey
+          );
+          url = appendToURLPath(extractedCreds.url, queueName);
+          options.proxy = extractedCreds.proxyUri;
+          pipeline = newPipeline(sharedKeyCredential, options);
+        } else {
+          throw new Error("Account connection string is only supported in Node.js environment");
+        }
+      } else if (extractedCreds.kind === "SASConnString") {
+        const queueName = credentialOrPipelineOrQueueName;
+        url = appendToURLPath(extractedCreds.url, queueName) + "?" + extractedCreds.accountSas;
+        pipeline = newPipeline(new AnonymousCredential(), options);
+      } else {
+        throw new Error(
+          "Connection string must be either an Account connection string or a SAS connection string"
+        );
+      }
+    } else {
+      throw new Error("Expecting non-empty strings for queueName parameter");
     }
     super(url, pipeline);
+    this._queueName = this.getQueueNameFromUrl();
     this.serviceContext = new Service(this.storageClientContext);
-  }
-
-  /**
-   * Creates a QueueClient object.
-   * @param queueName
-   */
-  public getQueueClient(queueName: string): QueueClient {
-    return new QueueClient(appendToURLPath(this.url, queueName), this.pipeline);
+    this.queueContext = new Queue(this.storageClientContext);
   }
 
   /**
@@ -257,11 +452,11 @@ export class QueueServiceClient extends StorageClient {
    * @returns {Promise<Models.ServiceGetPropertiesResponse>} Response data including the queue service properties.
    * @memberof QueueServiceClient
    */
-  public async getProperties(
+  public async getServiceProperties(
     options: ServiceGetPropertiesOptions = {}
   ): Promise<Models.ServiceGetPropertiesResponse> {
     const { span, spanOptions } = createSpan(
-      "QueueServiceClient-getProperties",
+      "QueueServiceClient-getServiceProperties",
       options.spanOptions
     );
     try {
@@ -290,12 +485,12 @@ export class QueueServiceClient extends StorageClient {
    * @returns {Promise<Models.ServiceSetPropertiesResponse>} Response data for the Set Properties operation.
    * @memberof QueueServiceClient
    */
-  public async setProperties(
+  public async setServiceProperties(
     properties: Models.QueueServiceProperties,
     options: ServiceGetPropertiesOptions = {}
   ): Promise<Models.ServiceSetPropertiesResponse> {
     const { span, spanOptions } = createSpan(
-      "QueueServiceClient-setProperties",
+      "QueueServiceClient-setServiceProperties",
       options.spanOptions
     );
     try {
@@ -324,11 +519,11 @@ export class QueueServiceClient extends StorageClient {
    * @returns {Promise<Models.ServiceGetStatisticsResponse>} Response data for get statistics the operation.
    * @memberof QueueServiceClient
    */
-  public async getStatistics(
+  public async getServiceStatistics(
     options: ServiceGetStatisticsOptions = {}
   ): Promise<Models.ServiceGetStatisticsResponse> {
     const { span, spanOptions } = createSpan(
-      "QueueServiceClient-getStatistics",
+      "QueueServiceClient-getServiceStatistics",
       options.spanOptions
     );
     try {
@@ -539,5 +734,256 @@ export class QueueServiceClient extends StorageClient {
         });
       }
     };
+  }
+
+  /**
+   * Creates a new queue under the specified account.
+   * @see https://docs.microsoft.com/en-us/rest/api/storageservices/create-queue4
+   *
+   * @param {QueueCreateOptions} [options] Options to Queue create operation.
+   * @returns {Promise<Models.QueueCreateResponse>} Response data for the Queue create operation.
+   * @memberof QueueServiceClient
+   */
+  public async create(options: QueueCreateOptions = {}): Promise<Models.QueueCreateResponse> {
+    const { span, spanOptions } = createSpan("QueueServiceClient-create", options.spanOptions);
+    try {
+      return this.queueContext.create({
+        ...options,
+        abortSignal: options.abortSignal,
+        spanOptions
+      });
+    } catch (e) {
+      span.setStatus({
+        code: CanonicalCode.UNKNOWN,
+        message: e.message
+      });
+      throw e;
+    } finally {
+      span.end();
+    }
+  }
+
+  /**
+   * Creates a QueueClient object.
+   * @param queueName
+   */
+  public getQueueClient(): QueueClient {
+    return new QueueClient(appendToURLPath(this.url, "messages"), this.pipeline);
+  }
+
+  /**
+   * Gets all user-defined metadata and system properties for the specified
+   * queue. Metadata is associated with the queue as name-values pairs.
+   * @see https://docs.microsoft.com/en-us/rest/api/storageservices/get-queue-metadata
+   *
+   * @param {QueueGetPropertiesOptions} [options] Options to Queue get properties operation.
+   * @returns {Promise<Models.QueueGetPropertiesResponse>} Response data for the Queue get properties operation.
+   * @memberof QueueServiceClient
+   */
+  public async getProperties(
+    options: QueueGetPropertiesOptions = {}
+  ): Promise<Models.QueueGetPropertiesResponse> {
+    const { span, spanOptions } = createSpan(
+      "QueueServiceClient-getProperties",
+      options.spanOptions
+    );
+    try {
+      return this.queueContext.getProperties({
+        abortSignal: options.abortSignal,
+        spanOptions
+      });
+    } catch (e) {
+      span.setStatus({
+        code: CanonicalCode.UNKNOWN,
+        message: e.message
+      });
+      throw e;
+    } finally {
+      span.end();
+    }
+  }
+
+  /**
+   * Deletes the specified queue permanently.
+   * @see https://docs.microsoft.com/en-us/rest/api/storageservices/delete-queue3
+   *
+   * @param {QueueDeleteOptions} [options] Options to Queue delete operation.
+   * @returns {Promise<Models.QueueDeleteResponse>} Response data for the Queue delete operation.
+   * @memberof QueueServiceClient
+   */
+  public async delete(options: QueueDeleteOptions = {}): Promise<Models.QueueDeleteResponse> {
+    const { span, spanOptions } = createSpan("QueueServiceClient-delete", options.spanOptions);
+    try {
+      return this.queueContext.deleteMethod({
+        abortSignal: options.abortSignal,
+        spanOptions
+      });
+    } catch (e) {
+      span.setStatus({
+        code: CanonicalCode.UNKNOWN,
+        message: e.message
+      });
+      throw e;
+    } finally {
+      span.end();
+    }
+  }
+
+  /**
+   * Sets one or more user-defined name-value pairs for the specified queue.
+   *
+   * If no option provided, or no metadata defined in the option parameter, the queue
+   * metadata will be removed.
+   * @see https://docs.microsoft.com/en-us/rest/api/storageservices/set-queue-metadata
+   *
+   * @param {Metadata} [metadata] If no metadata provided, all existing metadata will be removed.
+   * @param {QueueSetMetadataOptions} [options] Options to Queue set metadata operation.
+   * @returns {Promise<Models.QueueSetMetadataResponse>} Response data for the Queue set metadata operation.
+   * @memberof QueueServiceClient
+   */
+  public async setMetadata(
+    metadata?: Metadata,
+    options: QueueSetMetadataOptions = {}
+  ): Promise<Models.QueueSetMetadataResponse> {
+    const { span, spanOptions } = createSpan("QueueServiceClient-setMetadata", options.spanOptions);
+    try {
+      return this.queueContext.setMetadata({
+        abortSignal: options.abortSignal,
+        metadata,
+        spanOptions
+      });
+    } catch (e) {
+      span.setStatus({
+        code: CanonicalCode.UNKNOWN,
+        message: e.message
+      });
+      throw e;
+    } finally {
+      span.end();
+    }
+  }
+
+  /**
+   * Gets details about any stored access policies specified on the queue that may be used with Shared Access Signatures.
+   *
+   * WARNING: JavaScript Date will potential lost precision when parsing start and expiry string.
+   * For example, new Date("2018-12-31T03:44:23.8827891Z").toISOString() will get "2018-12-31T03:44:23.882Z".
+   *
+   * @see https://docs.microsoft.com/en-us/rest/api/storageservices/get-queue-acl
+   *
+   * @param {QueueGetAccessPolicyOptions} [options] Options to Queue get access policy operation.
+   * @returns {Promise<QueueGetAccessPolicyResponse>} Response data for the Queue get access policy operation.
+   * @memberof QueueServiceClient
+   */
+  public async getAccessPolicy(
+    options: QueueGetAccessPolicyOptions = {}
+  ): Promise<QueueGetAccessPolicyResponse> {
+    const { span, spanOptions } = createSpan(
+      "QueueServiceClient-getAccessPolicy",
+      options.spanOptions
+    );
+    try {
+      const response = await this.queueContext.getAccessPolicy({
+        abortSignal: options.abortSignal,
+        spanOptions
+      });
+
+      const res: QueueGetAccessPolicyResponse = {
+        _response: response._response,
+        date: response.date,
+        requestId: response.requestId,
+        clientRequestId: response.clientRequestId,
+        signedIdentifiers: [],
+        version: response.version,
+        errorCode: response.errorCode
+      };
+
+      for (const identifier of response) {
+        res.signedIdentifiers.push({
+          accessPolicy: {
+            expiry: new Date(identifier.accessPolicy.expiry),
+            permission: identifier.accessPolicy.permission,
+            start: new Date(identifier.accessPolicy.start)
+          },
+          id: identifier.id
+        });
+      }
+
+      return res;
+    } catch (e) {
+      span.setStatus({
+        code: CanonicalCode.UNKNOWN,
+        message: e.message
+      });
+      throw e;
+    } finally {
+      span.end();
+    }
+  }
+
+  /**
+   * Sets stored access policies for the queue that may be used with Shared Access Signatures.
+   * @see https://docs.microsoft.com/en-us/rest/api/storageservices/set-queue-acl
+   *
+   * @param {PublicAccessType} [access]
+   * @param {SignedIdentifier[]} [queueAcl]
+   * @param {QueueSetAccessPolicyOptions} [options] Options to Queue set access policy operation.
+   * @returns {Promise<Models.QueueSetAccessPolicyResponse>} Response data for the Queue set access policy operation.
+   * @memberof QueueServiceClient
+   */
+  public async setAccessPolicy(
+    queueAcl?: SignedIdentifier[],
+    options: QueueSetAccessPolicyOptions = {}
+  ): Promise<Models.QueueSetAccessPolicyResponse> {
+    const { span, spanOptions } = createSpan(
+      "QueueServiceClient-setAccessPolicy",
+      options.spanOptions
+    );
+    try {
+      const acl: Models.SignedIdentifier[] = [];
+      for (const identifier of queueAcl || []) {
+        acl.push({
+          accessPolicy: {
+            expiry: truncatedISO8061Date(identifier.accessPolicy.expiry),
+            permission: identifier.accessPolicy.permission,
+            start: truncatedISO8061Date(identifier.accessPolicy.start)
+          },
+          id: identifier.id
+        });
+      }
+
+      return this.queueContext.setAccessPolicy({
+        abortSignal: options.abortSignal,
+        queueAcl: acl,
+        spanOptions
+      });
+    } catch (e) {
+      span.setStatus({
+        code: CanonicalCode.UNKNOWN,
+        message: e.message
+      });
+      throw e;
+    } finally {
+      span.end();
+    }
+  }
+
+  private getQueueNameFromUrl(): string {
+    //  URL may look like the following
+    // "https://myaccount.queue.core.windows.net/myqueue?sasString".
+    // "https://myaccount.queue.core.windows.net/myqueue".
+    try {
+      let urlWithoutSAS = this.url.split("?")[0]; // removing the sas part of url if present
+      urlWithoutSAS = urlWithoutSAS.endsWith("/") ? urlWithoutSAS.slice(0, -1) : urlWithoutSAS; // Slicing off '/' at the end if exists
+
+      const queueName = urlWithoutSAS.match("([^/]*)://([^/]*)/([^/]*)")![3];
+      if (!queueName) {
+        throw new Error("Provided queueName is invalid.");
+      } else {
+        return queueName;
+      }
+    } catch (error) {
+      throw new Error("Unable to extract queueName with provided information.");
+    }
   }
 }
