+++ sdk/storage/storage-queue/src/index.browser.ts
@@ -8,8 +8,6 @@ import * as Models from "./generated/src/models";
 export * from "./credentials/AnonymousCredential";
 export * from "./credentials/Credential";
 export { IPRange } from "./IPRange";
-export * from "./MessageIdClient";
-export * from "./MessagesClient";
 export * from "./Pipeline";
 export * from "./policies/AnonymousCredentialPolicy";
 export * from "./policies/CredentialPolicy";
